requirements.txt
nats.py==2.2.0
nats.py[nkeys]==0.1.0

